requirements.txt
Django>=4.0
djangorestframework>=3.13
django-cors-headers>=3.13
whitenoise>=6.0
Pillow>=9.0
django-ckeditor>=6.5.1
gunicorn>=20.1.0